Refactor runners to share setup
