+++ Runner/Automation.Runner.csproj
@@ -1,21 +1,14 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-	  <TargetFramework>net8.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="10.0.0-preview.3.25171.5" />
   </ItemGroup>
-
   <ItemGroup>
-    <ProjectReference Include="..\src\Automation.Core\Automation.Core.csproj" />
-    <ProjectReference Include="..\src\Automation.Engines\Automation.Engine.CDP\Automation.Engine.CDP.csproj" />
-    <ProjectReference Include="..\src\Automation.Engines\Automation.Engine.UIAutomation\Automation.Engine.UIAutomation.csproj" />
-    <ProjectReference Include="..\src\Automation.Tasks\Automation.Tasks.csproj" />
+    <ProjectReference Include="..\src\Automation.Runtime\Automation.Runtime.csproj" />
   </ItemGroup>
-
 </Project>
